refactor: alert_texts/ を feat-recruit/common/alert_texts/ に移動し、import パスを更新
