+++ src/app/feat-recruit/close_recruit/auto_close.ts
@@ -4,10 +4,10 @@ import { ParticipantService } from '../../../db/participant_service';
 import { RecruitService } from '../../../db/recruit_service';
 import { setButtonDisable } from '../../common/button_components';
 import { notExists } from '../../common/others';
-import { regenerateCanvas, RecruitOpCode } from '../canvases/regenerate_canvas';
+import { regenerateCanvas, RecruitOpCode } from '../common/canvases/regenerate_canvas';
+import { RecruitData } from '../common/types/recruit_data';
 import { getMemberMentions } from '../interactions/buttons/other_events';
 import { sendCloseEmbedSticky } from '../sticky/recruit_sticky_messages';
-import { RecruitData } from '../types/recruit_data';
 
 export async function recruitAutoClose(
     recruitData: RecruitData,
