+++ src/app/feat-recruit/common/alert_texts/alert_texts.ts
