+++ src/app/feat-recruit/common/alert_texts/schedule_related_alerts.ts
@@ -1,5 +1,5 @@
-import { UniqueChannelService } from '../../../db/unique_channel_service';
-import { ChannelKeySet } from '../../constant/channel_key';
+import { UniqueChannelService } from '../../../../db/unique_channel_service';
+import { ChannelKeySet } from '../../../constant/channel_key';
 
 export async function getFestPeriodAlertText(guildId: string) {
     const shiverChannelId = await UniqueChannelService.getChannelIdByKey(
