+++ src/app/feat-recruit/create_recruit/common/remove_delete_button.ts
@@ -1,7 +1,7 @@
 import { log4js_obj } from '../../../../log4js_settings';
 import { searchMessageById } from '../../../common/manager/message_manager';
 import { exists } from '../../../common/others';
-import { RecruitData } from '../../types/recruit_data';
+import { RecruitData } from '../../common/types/recruit_data';
 
 const logger = log4js_obj.getLogger('recruit');
