+++ src/app/feat-recruit/create_recruit/condition_checks/schedule_check.ts
@@ -8,8 +8,8 @@ import {
 } from '../../../common/apis/splatoon3.ink/splatoon3_ink';
 import { Sp3Schedule } from '../../../common/apis/splatoon3.ink/types/schedule';
 import { notExists } from '../../../common/others';
-import { RecruitAlertTexts } from '../../alert_texts/alert_texts';
-import { getFestPeriodAlertText } from '../../alert_texts/schedule_related_alerts';
+import { RecruitAlertTexts } from '../../common/alert_texts/alert_texts';
+import { getFestPeriodAlertText } from '../../common/alert_texts/schedule_related_alerts';
 
 type checkRecruitScheduleResponse = {
     canRecruit: boolean;
