+++ src/app/feat-recruit/edit_recruit/recruit_edit.ts
@@ -10,8 +10,8 @@ import { ErrorTexts } from '../../constant/error_texts';
 import { StickyKey } from '../../constant/sticky_key';
 import { sendErrorLogs } from '../../logs/error/send_error_logs';
 import { sendEditRecruitLog } from '../../logs/modals/recruit_modal_log';
-import { RecruitOpCode, regenerateCanvas } from '../canvases/regenerate_canvas';
-import { regenerateEmbed } from '../embeds/regenerate_embed';
+import { RecruitOpCode, regenerateCanvas } from '../common/canvases/regenerate_canvas';
+import { regenerateEmbed } from '../common/embeds/regenerate_embed';
 import { availableRecruitString } from '../sticky/recruit_sticky_messages';
 
 const logger = log4js_obj.getLogger('interaction');
